FALTA FUNÇÕES DA UNIDADE TRABALHO

+++ src/main/java/br/com/alura/spring/data/services/CrudCargoService.java
@@ -1,11 +1,15 @@
 package br.com.alura.spring.data.services;
 
 import br.com.alura.spring.data.entities.Cargo;
+import br.com.alura.spring.data.exception.ValidacaoException;
 import br.com.alura.spring.data.repository.CargoRepository;
 import br.com.alura.spring.data.repository.FuncionarioRepository;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+import java.util.Optional;
 import java.util.Scanner;
+import java.util.ArrayList;
 
 @Service
 public class CrudCargoService {
@@ -21,6 +25,7 @@ public class CrudCargoService {
 
     public void inicial(){
         System.out.println("Bem vindo ao cadastro de Cargos:");
+
         while(controleFluxoLacoRepeticao){
             System.out.println("0.Sair \n1.Cadastrar Cargo \n2.Atualizar Cargo \n3. Visualizar Todos o Cargos \n4.Deletar Cargo Por Id");
             int opcao = sc.nextInt();
@@ -41,19 +46,9 @@ public class CrudCargoService {
                     controleFluxoLacoRepeticao = false;
                     break;
             }
-            if(controleFluxoLacoRepeticao) {
-                System.out.println("Deseja sair: \n 1.Sim \n 2.Não");
-                int ficarOuSair = sc.nextInt();
-                if (ficarOuSair == 2) {
-                    controleFluxoLacoRepeticao = true;
-                } else {
-                    controleFluxoLacoRepeticao = false;
-                    System.out.println("Obrigado por utilizar os melhores recursos em tecnologia!");
-                }
-            }
-            else {
-                System.out.println("Obrigado por utilizar os melhores recursos em tecnologia!");
-            }
+            System.out.println("Deseja continuar? \n 1.Sim \n 2.Não");
+            Integer desejaContinuar = sc.nextInt();
+            controleFluxoLacoRepeticao = (desejaContinuar == 1) ? true : false;
         }
 
 
@@ -62,9 +57,11 @@ public class CrudCargoService {
 
     private void salvar(){
 
-        System.out.println("Digite o nome do cargo");
+        System.out.print("Digite o nome do cargo: ");
+        sc.nextLine();
         String nomeCargo = sc.next();
         Cargo cargo = new Cargo(nomeCargo);
+
         cargoRepository.save(cargo);
         System.out.println("Cargo cadastrado com sucesso");
     }
@@ -72,11 +69,16 @@ public class CrudCargoService {
     private void atualizar() {
         System.out.println("Digite o número do ID que deseja recuperar: ");
         Integer idRecuperado = sc.nextInt();
-        System.out.println("Digite o novo nome do cadastro");
-        String nomeParaAtualizar = sc.next();
 
-        Cargo cargo = new Cargo(nomeParaAtualizar);
-        cargo.setId(idRecuperado);
+        Optional<Cargo> cargoRecuperado = cargoRepository.findById(idRecuperado);
+
+        if (cargoRecuperado.isEmpty()){
+            throw new ValidacaoException("ERROR! Cargo não encontrado");
+        }
+        System.out.print("Digite o novo nome do cadastro: ");
+        sc.nextLine();
+        String nomeParaAtualizar = sc.nextLine();
+        Cargo cargo = new Cargo(idRecuperado,nomeParaAtualizar);
 
         cargoRepository.save(cargo);
 
@@ -91,7 +93,16 @@ public class CrudCargoService {
     private void deletarPorId(){
         System.out.println("Digite o Id que deseja excluir do banco de dados");
         Integer idParaDeletar = sc.nextInt();
+
+        Optional<Cargo> cargoRecuperado = cargoRepository.findById(idParaDeletar);
+
+        if (cargoRecuperado.isEmpty()){
+            throw new ValidacaoException("ERROR! Cargo não encontrado");
+        }
+
         cargoRepository.deleteById(idParaDeletar);
         System.out.println("Registro Deletado Com Sucesso");
     }
+
+
 }
